cleanup: move files to correct place
